+++ client/src/components/Notifiction/NotifiactionData.jsx
@@ -2,14 +2,31 @@ import React from "react";
 import { deleteOneNotifications } from "../../../store (3)/api/notificationApi";
 import { useDispatch } from "react-redux";
 import { FaTrash, FaUser } from "react-icons/fa";
-import { BsPostcardHeart } from "react-icons/bs";
 import { FaHeart } from "react-icons/fa6";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 function NotifiactionData({ notifiction }) {
   const dispatch = useDispatch();
   const handleDeleteOneNotifiaction = (id) => {
-    dispatch(deleteOneNotifications(id));
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't delete this notifiction!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteOneNotifications(id));
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success",
+        });
+      }
+    });
   };
   const handleType = () => {
     if (notifiction.type === "follow") {
